requirements.txt
pre-commit
trimesh
torch>=2
torchvision
einops
huggingface_hub
imageio
numpy<2
opencv-python
xformers
open3d
fastapi
unicorn
requests
typer
pillow
omegaconf
evo
e3nn
moviepy==1.0.3
plyfile
pillow_heif
safetensors
websockets>=11.0
scipy
addict

